OUT OF ORDER
Yes, the commit dates appear out of order relative to the repository tree history.
The author and commit dates represent the original author and publication dates for the content posted.
The repository structure was revised to best represent the relative order of content creation and
contributions to this repository, while also preserving the project file maintenance and configuration history.

This required a non-linear repository branch structure to preserve the commit history.